pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Denis Guryanov <guryanov19962010@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.1"
python-dotenv = "^1.1.0"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
pillow = "^11.2.1"
black = "^25.1.0"
isort = "^6.0.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.1.0"}
django-countries = "^7.6.1"
django-filter = "^25.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
rest-framework-simplejwt = "^0.0.2"
coverage = "^7.8.2"
drf-yasg = "^1.21.10"
stripe = "^12.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
